requirements.txt
fastapi
uvicorn
pydantic>=2.0.0
openai
python-multipart
jinja2
pyyaml
python-dotenv
requests
colorama
